Exceções

+++ src/main/java/com/example/valeriapaz/ecommerce/services/ProductService.java
@@ -3,13 +3,16 @@ package com.example.valeriapaz.ecommerce.services;
 import com.example.valeriapaz.ecommerce.dto.ProductDTO;
 import com.example.valeriapaz.ecommerce.entities.Product;
 import com.example.valeriapaz.ecommerce.repositories.ProductRepository;
+import com.example.valeriapaz.ecommerce.services.exceptions.DatabaseException;
+import com.example.valeriapaz.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -18,20 +21,21 @@ public class ProductService {
     private ProductRepository repository;
 
     // Escrita didática
-    @Transactional(readOnly = true)
-    public ProductDTO findById(Long id) {
-        // Find by Id retorna um objeto do tipo Optional por padrão
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
-        return dto;
-    }
-    // Escrita mais curta!!
 //    @Transactional(readOnly = true)
 //    public ProductDTO findById(Long id) {
-//        Product product = repository.findById(id).get();
-//        return new ProductDTO(product);
+//        // Find by Id retorna um objeto do tipo Optional por padrão
+//        Optional<Product> result = repository.findById(id);
+//        Product product = result.get();
+//        ProductDTO dto = new ProductDTO(product);
+//        return dto;
 //    }
+    // Escrita mais curta!!
+    @Transactional(readOnly = true)
+    public ProductDTO findById(Long id) {
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
+        return new ProductDTO(product);
+    }
 
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable) {
@@ -49,15 +53,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
+
     }
